@@ -3,7 +3,7 @@ name = "apis"
 version = "0.1.0"
 description = "Various API interactions"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = [
     "beautifulsoup4>=4.14.2",
     "requests>=2.32.5",
